+++ src/style/height_mode.rs
@@ -5,12 +5,14 @@
 //! shrink the text box. Height modes help us achieve this.
 //!
 //! [`TextBox`]: ../../struct.TextBox.html
-use embedded_graphics::prelude::*;
-
+use crate::style::vertical_overdraw::VerticalOverdraw;
 use crate::{
     alignment::{HorizontalTextAlignment, VerticalTextAlignment},
+    rendering::cursor::Cursor,
     StyledTextBox,
 };
+use core::ops::Range;
+use embedded_graphics::prelude::*;
 
 /// Specifies how the [`TextBox`]'s height is adjusted when it is turned into a [`StyledTextBox`].
 ///
@@ -29,6 +31,13 @@ pub trait HeightMode: Copy {
         A: HorizontalTextAlignment,
         V: VerticalTextAlignment,
         H: HeightMode;
+
+    /// Calculate the range of rows of the current line that can be drawn.
+    ///
+    /// If a line does not fully fit in the bounding box, some `HeightMode` options allow drawing
+    /// partial lines. For a partial line, this function calculates, which rows of each character
+    /// should be displayed.
+    fn calculate_displayed_row_range<F: Font>(cursor: &Cursor<F>) -> Range<i32>;
 }
 
 /// Keep the original [`TextBox`] height.
@@ -58,9 +67,12 @@ pub trait HeightMode: Copy {
 ///
 /// [`TextBox`]: ../../struct.TextBox.html
 #[derive(Copy, Clone, Debug)]
-pub struct Exact;
+pub struct Exact<OV: VerticalOverdraw>(pub OV);
 
-impl HeightMode for Exact {
+impl<OV> HeightMode for Exact<OV>
+where
+    OV: VerticalOverdraw,
+{
     #[inline]
     fn apply<C, F, A, V, H>(_text_box: &mut StyledTextBox<'_, C, F, A, V, H>)
     where
@@ -70,6 +82,11 @@ impl HeightMode for Exact {
         V: VerticalTextAlignment,
         H: HeightMode,
     {
+    }
+
+    #[inline]
+    fn calculate_displayed_row_range<F: Font>(cursor: &Cursor<F>) -> Range<i32> {
+        OV::calculate_displayed_row_range(cursor)
     }
 }
 
@@ -141,6 +158,13 @@ impl HeightMode for FitToText {
     {
         text_box.fit_height();
     }
+
+    #[inline]
+    fn calculate_displayed_row_range<F: Font>(_: &Cursor<F>) -> Range<i32> {
+        // FitToText always sets the bounding box to the exact size of the text, so every row is
+        // always fully displayed
+        0..F::CHARACTER_SIZE.height as i32
+    }
 }
 
 /// If the text does not fill the bounding box, shrink the [`StyledTextBox`] to be as tall as the
@@ -149,7 +173,7 @@ impl HeightMode for FitToText {
 /// # Example: `ShrinkToText` does not grow the [`TextBox`].
 ///
 /// ```rust
-/// use embedded_text::prelude::*;
+/// use embedded_text::{prelude::*, style::vertical_overdraw::FullRowsOnly};
 /// use embedded_graphics::{fonts::Font6x8, pixelcolor::BinaryColor, prelude::*};
 ///
 /// // This TextBox contains two lines of text, but is 1px high
@@ -160,7 +184,7 @@ impl HeightMode for FitToText {
 ///
 /// // Set style, use 6x8 font so the 2 lines are 16px high.
 /// let style = TextBoxStyleBuilder::new(Font6x8)
-///     .height_mode(ShrinkToText)
+///     .height_mode(ShrinkToText(FullRowsOnly))
 ///     .text_color(BinaryColor::On)
 ///     .build();
 ///
@@ -171,7 +195,7 @@ impl HeightMode for FitToText {
 /// # Example: `ShrinkToText` shrinks the [`TextBox`].
 ///
 /// ```rust
-/// use embedded_text::prelude::*;
+/// use embedded_text::{prelude::*, style::vertical_overdraw::FullRowsOnly};
 /// use embedded_graphics::{fonts::Font6x8, pixelcolor::BinaryColor, prelude::*};
 ///
 /// // This TextBox contains two lines of text, but is 60px high
@@ -182,7 +206,7 @@ impl HeightMode for FitToText {
 ///
 /// // Set style, use 6x8 font so the 2 lines are 16px high.
 /// let style = TextBoxStyleBuilder::new(Font6x8)
-///     .height_mode(ShrinkToText)
+///     .height_mode(ShrinkToText(FullRowsOnly))
 ///     .text_color(BinaryColor::On)
 ///     .build();
 ///
@@ -192,9 +216,12 @@ impl HeightMode for FitToText {
 ///
 /// [`TextBox`]: ../../struct.TextBox.html
 #[derive(Copy, Clone, Debug)]
-pub struct ShrinkToText;
+pub struct ShrinkToText<OV: VerticalOverdraw>(pub OV);
 
-impl HeightMode for ShrinkToText {
+impl<OV> HeightMode for ShrinkToText<OV>
+where
+    OV: VerticalOverdraw,
+{
     #[inline]
     fn apply<C, F, A, V, H>(text_box: &mut StyledTextBox<'_, C, F, A, V, H>)
     where
@@ -205,5 +232,10 @@ impl HeightMode for ShrinkToText {
         H: HeightMode,
     {
         text_box.fit_height_limited(text_box.size().height);
+    }
+
+    #[inline]
+    fn calculate_displayed_row_range<F: Font>(cursor: &Cursor<F>) -> Range<i32> {
+        OV::calculate_displayed_row_range(cursor)
     }
 }