+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
  * `TextBoxStyleBuilder::from_text_style`
  * Added `HeightMode` to select whether and how the `StyledTextBox` height should be aligned to the
    actual text height.
+ * Added `Hidden`, `Visible` and `FullRowsOnly` overflow control modes to `Exact` and `ShrinkToText` height modes.
 
 ## Changed: